Move flow service into flows module
